requirements.txt
pytelegrambotapi==2.3.0
jinja2==2.8
flask==0.11.1
Flask-SQLAlchemy==2.1
Flask-Admin-1.4.2